Reorganize `NettyBufferProvider`, `NettyByteBuf` and improve internal docs
